WIP. Fixed warning, casting int into uint16_t.

+++ main-program/TestCookieDough/DataPacketHelperTest.cpp
@@ -12,7 +12,7 @@ void Helper_ParsePacketFromEEPROM_PrepareToReturnPacket_ParsePacketFromEepromSuc
     Helper_ParsePacketFromEEPROM_PrepareToReturnPacket(packet);
     DataPacket *resultPtr = new DataPacket(); 
     DataPacket result = *resultPtr;
-    unsigned int packetSize;
+    uint16_t packetSize;
     bool resultBool = ParsePacketFromEEPROM(0, &result, &packetSize);
 
     ASSERT_TEST(resultBool == true &&
