+++ main-program/TestCookieDough/DataPacketTest.cpp
@@ -109,10 +109,10 @@ void TestIfVectorTestsAreWorking_ShouldReturnDifferentParametersEachTime()
     EEPROM.read(2);
     EEPROM.read(3);
     EEPROM.read(4);
-    uint8_t first = EEPROMClass_read_param_idx_v[0];
-    uint8_t second = EEPROMClass_read_param_idx_v[1];
-    uint8_t third = EEPROMClass_read_param_idx_v[2];
-    uint8_t fourth = EEPROMClass_read_param_idx_v[3];
+    int first = EEPROMClass_read_param_idx_v[0];
+    int second = EEPROMClass_read_param_idx_v[1];
+    int third = EEPROMClass_read_param_idx_v[2];
+    int fourth = EEPROMClass_read_param_idx_v[3];
 
     ASSERT_TEST(first == 1 && second == 2 && third == 3 && fourth == 4);
 }
@@ -168,8 +168,8 @@ void TestIfOverwriteReferenceResettingIsWorking_ResetsBetweenTests()
 void SavePacketToEEPROM_SavesStxToFirstGivenAdress()
 {
     uint8_t data = 42;
-    unsigned int adress = 20;
-    unsigned int packetSize;
+    uint16_t adress = 20;
+    uint16_t packetSize;
 
     SavePacketToEEPROM(adress, &data, sizeof(data), &packetSize);
 
@@ -180,8 +180,8 @@ void SavePacketToEEPROM_EtxIsPutDownAtTheEndOfThePacket()
 {
     DataPacket packet;
     uint8_t data = 42;
-    unsigned int adress = 20;
-    unsigned int packetSize;
+    uint16_t adress = 20;
+    uint16_t packetSize;
 
     SavePacketToEEPROM(adress, &data, sizeof(data), &packetSize);
     int expectedEtxPosition = adress + sizeof(packet.etx) + sizeof(packet.payloadLength) + sizeof(packet.crc) + sizeof(data);
@@ -193,7 +193,7 @@ void SavePacketToEEPROM_PacketIsCorrectlyPutDown()
 {
     uint16_t data = 42;
     uint8_t *dataPtr = (uint8_t*) &data;
-    unsigned int adress = 20;
+    uint16_t adress = 20;
     DataPacket packet;
     int expectedStxAdress = static_cast<int>(adress);
     int expectedPayloadLengthAdress = adress + sizeof(packet.stx);
@@ -211,7 +211,7 @@ void SavePacketToEEPROM_PacketIsCorrectlyPutDown()
     EEPROMClass_read_return_v.push_back(dataPtr[0]);
     EEPROMClass_read_return_v.push_back(dataPtr[1]);
 
-    unsigned int packetSize;
+    uint16_t packetSize;
     bool resultBool = SavePacketToEEPROM(adress, dataPtr, sizeof(data), &packetSize);
 
     ASSERT_TEST(resultBool == true && 
@@ -228,7 +228,7 @@ void SavePacketToEEPROM_PacketIsSavedButEepromFailsToReadTheData_ReturnsFalse()
 {
     uint16_t data = 42;
     uint8_t *dataPtr = (uint8_t*) &data;
-    unsigned int adress = 20;
+    uint16_t adress = 20;
     DataPacket packet;
 
     // This ensures that ParsePacketFromEEPROM returns true
@@ -237,10 +237,10 @@ void SavePacketToEEPROM_PacketIsSavedButEepromFailsToReadTheData_ReturnsFalse()
     EEPROMClass_length_return = sizeof(data) + 20;
     EEPROMClass_get_param_t_o2_vr.push_back(55561893);
     EEPROMClass_read_return_v.push_back(packet.etx);
-    EEPROMClass_read_return_v.push_back(dataPtr[0] + 13);
+    EEPROMClass_read_return_v.push_back(static_cast<uint8_t>(dataPtr[0] + 13));
     EEPROMClass_read_return_v.push_back(dataPtr[1]);
 
-    unsigned int packetSize;
+    uint16_t packetSize;
     bool resultBool = SavePacketToEEPROM(adress, dataPtr, sizeof(data), &packetSize);
 
     ASSERT_TEST(resultBool == false);
@@ -250,8 +250,8 @@ void SavePacketToEEPROM_AdaptsSizeOfPacketToFitData()
 {
     DataPacket packet;
     uint16_t data = 42;
-    unsigned int adress = 20;
-    unsigned int packetSize;
+    uint16_t adress = 20;
+    uint16_t packetSize;
 
     int expectedPayloadAdress = adress + sizeof(packet.stx) + sizeof(packet.payloadLength) + sizeof(packet.crc);
 
@@ -263,7 +263,7 @@ void SavePacketToEEPROM_AdaptsSizeOfPacketToFitData()
 
 void ParsePacketFromEEPROM_ReturnsCorrectPackage() // TODO: bad test. Locks the order the mocked functions will be called. Rewrite with different framework?
 {
-    unsigned int adress = 13;
+    uint16_t adress = 13;
     uint16_t data = 421;
     DataPacket expectedPacket;
     expectedPacket.payloadLength = sizeof(data);
@@ -272,7 +272,7 @@ void ParsePacketFromEEPROM_ReturnsCorrectPackage() // TODO: bad test. Locks the 
 
     EEPROMClass_read_return_v.push_back(expectedPacket.stx);
     EEPROMClass_get_param_t_o1_vr.push_back(expectedPacket.payloadLength);
-    EEPROMClass_length_return = expectedPacket.payloadLength + 20;
+    EEPROMClass_length_return = static_cast<uint16_t>(expectedPacket.payloadLength + 20);
     EEPROMClass_get_param_t_o2_vr.push_back(expectedPacket.crc);
     EEPROMClass_read_return_v.push_back(expectedPacket.etx);
     EEPROMClass_read_return_v.push_back(expectedPacket.payload[0]);
@@ -280,7 +280,7 @@ void ParsePacketFromEEPROM_ReturnsCorrectPackage() // TODO: bad test. Locks the 
 
     DataPacket *resultPtr = new DataPacket(); 
     DataPacket result = *resultPtr;
-    unsigned int packetSize;
+    uint16_t packetSize;
     bool resultBool = ParsePacketFromEEPROM(adress, &result, &packetSize);
 
     ASSERT_TEST(resultBool == true &&
@@ -296,7 +296,7 @@ void ParsePacketFromEEPROM_ReturnsCorrectPackage() // TODO: bad test. Locks the 
 
 void ParsePacketFromEEPROM_EepromReturnsFaultyData_ReturnsFalse() // TODO: bad test. Locks the order the mocked functions will be called. Rewrite with different framework?
 {
-    unsigned int adress = 13;
+    uint16_t adress = 13;
     uint16_t data = 421;
     DataPacket expectedPacket;
     expectedPacket.payloadLength = sizeof(data);
@@ -305,15 +305,15 @@ void ParsePacketFromEEPROM_EepromReturnsFaultyData_ReturnsFalse() // TODO: bad t
 
     EEPROMClass_read_return_v.push_back(expectedPacket.stx);
     EEPROMClass_get_param_t_o1_vr.push_back(expectedPacket.payloadLength);
-    EEPROMClass_length_return = expectedPacket.payloadLength + 20;
+    EEPROMClass_length_return = static_cast<uint16_t>(expectedPacket.payloadLength + 20);
     EEPROMClass_get_param_t_o2_vr.push_back(expectedPacket.crc);
     EEPROMClass_read_return_v.push_back(expectedPacket.etx);
-    EEPROMClass_read_return_v.push_back(expectedPacket.payload[0] + 13);
+    EEPROMClass_read_return_v.push_back(static_cast<uint8_t>(expectedPacket.payload[0] + 13));
     EEPROMClass_read_return_v.push_back(expectedPacket.payload[1]);
 
     DataPacket *resultPtr = new DataPacket(); 
     DataPacket result = *resultPtr;
-    unsigned int packetSize;
+    uint16_t packetSize;
     bool resultBool = ParsePacketFromEEPROM(adress, &result, &packetSize);
 
     ASSERT_TEST(resultBool == false);
